React Flow might actually be a bad idea, I've recreated what we have in vinilla d3, now going to port it to react
